Move examples to a separate folder
Also GridNodeUI is moved to its example file because it is not worthy to
be a stock Node UI type.
