+++ src/DynodeUI/index.ts
@@ -1,11 +1,9 @@
 import StageUI from './standardStageUI';
 import NodeUI from './nodeUI';
 import LinkUI from './linkUI';
-import GridNodeUI from './gridNodeUI';
 
 export {
   StageUI,
   NodeUI,
   LinkUI,
-  GridNodeUI,
 };
